server/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
cssselect==1.2.0
cssutils==2.11.1
databases==0.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.114.2
greenlet==3.1.0
h11==0.14.0
idna==3.10
lxml==5.3.0
more-itertools==10.5.0
passlib==1.7.4
pillow==10.4.0
premailer==3.10.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
rsa==4.9
setuptools==75.1.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
sqlmodel==0.0.22
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
wheel==0.44.0
yagmail==0.15.293
